SideBarWidget merged
