+++ web-ui/src/main/java/org/archcnl/ui/outputview/sidebar/SideBarWidget.java
@@ -5,6 +5,7 @@ import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.Text;
 import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
 import com.vaadin.flow.component.orderedlayout.VerticalLayout;
@@ -24,9 +25,9 @@ public class SideBarWidget extends VerticalLayout {
     private SideBarTab freeTextQueryTab;
 
     public SideBarWidget(
-            QueryResultsUiComponent defaultQueryView,
-            CustomQueryView customQueryView,
-            FreeTextQueryUiComponent freeTextQueryView) {
+            final QueryResultsUiComponent defaultQueryView,
+            final CustomQueryView customQueryView,
+            final FreeTextQueryUiComponent freeTextQueryView) {
         setHeightFull();
         getStyle().set("overflow", "hidden");
 
@@ -36,7 +37,7 @@ public class SideBarWidget extends VerticalLayout {
     }
 
     public void addReturnButton() {
-        Button returnButton =
+        final Button returnButton =
                 new Button(
                         "Return to rule editor",
                         new Icon(VaadinIcon.CHEVRON_LEFT),
@@ -46,26 +47,28 @@ public class SideBarWidget extends VerticalLayout {
     }
 
     public void addTabs(
-            QueryResultsUiComponent defaultQueryView,
-            CustomQueryView customQueryView,
-            FreeTextQueryUiComponent freeTextQueryView) {
+            final QueryResultsUiComponent defaultQueryView,
+            final CustomQueryView customQueryView,
+            final FreeTextQueryUiComponent freeTextQueryView) {
 
-        customQueryTab = new SideBarTab("Custom Queries", VaadinIcon.AUTOMATION, customQueryView);
-        freeTextQueryTab =
-                new SideBarTab("Free Text Queries", VaadinIcon.TEXT_INPUT, freeTextQueryView);
-        tabs =
-                new Tabs(
-                        new SideBarTab(
-                                "General Information", VaadinIcon.INFO_CIRCLE, defaultQueryView),
-                        customQueryTab,
-                        freeTextQueryTab);
+        final Span genInfo = new Span("General Information");
+        genInfo.setClassName("side-bar-text");
+        final SideBarTab generalInformationTab =
+                new SideBarTab(genInfo, VaadinIcon.INFO_CIRCLE, defaultQueryView);
+        final Span customQuery = new Span("Custom Queries");
+        customQuery.setClassName("side-bar-text");
+        customQueryTab = new SideBarTab(customQuery, VaadinIcon.AUTOMATION, customQueryView);
+        final Span freeText = new Span("Free Text Queries");
+        freeText.setClassName("side-bar-text");
+        freeTextQueryTab = new SideBarTab(freeText, VaadinIcon.TEXT_INPUT, freeTextQueryView);
+        tabs = new Tabs(generalInformationTab, customQueryTab, freeTextQueryTab);
         tabs.setOrientation(Tabs.Orientation.VERTICAL);
         tabs.setHeightFull();
 
         tabs.addSelectedChangeListener(
                 event -> {
                     if (event.getSelectedTab() instanceof SideBarTab) {
-                        SideBarTab tab = (SideBarTab) event.getSelectedTab();
+                        final SideBarTab tab = (SideBarTab) event.getSelectedTab();
                         fireEvent(
                                 new ShowComponentRequestedEvent(
                                         this, true, tab.getLinkedComponent()));
@@ -76,16 +79,18 @@ public class SideBarWidget extends VerticalLayout {
     }
 
     public void addPinnedQueryTab(Component linkedComponent, String name) {
-        SideBarTab newTab = new SideBarTab(name, VaadinIcon.PIN, linkedComponent);
+        final Span nameSpan = new Span(name);
+        nameSpan.setClassName("side-bar-text");
+        final SideBarTab newTab = new SideBarTab(nameSpan, VaadinIcon.PIN, linkedComponent);
         tabs.add(newTab);
         tabs.setSelectedTab(newTab);
     }
 
-    public void updateCustomQueryTab(CustomQueryView customQueryView) {
+    public void updateCustomQueryTab(final CustomQueryView customQueryView) {
         customQueryTab.setLinkedComponent(customQueryView);
     }
 
-    public void updateFreeTextQueryTab(FreeTextQueryUiComponent freeTextQueryView) {
+    public void updateFreeTextQueryTab(final FreeTextQueryUiComponent freeTextQueryView) {
         freeTextQueryTab.setLinkedComponent(freeTextQueryView);
     }
